Moved obsolete metric file
